Refactor: split View.swift into focused view files under Views/
Extracts IntervalsView, SettingsView, SoundsView, TBPopoverView, and
DayPickerView into separate files for better separation of concerns.
DayPickerView also gains Calendar injection for testability.
