+++ TomatoBar/Views/IntervalsView.swift
@@ -1,0 +1,76 @@
+import KeyboardShortcuts
+import LaunchAtLogin
+import SwiftUI
+
+extension KeyboardShortcuts.Name {
+    static let startStopTimer = Self("startStopTimer")
+}
+
+private enum ChildView {
+    case intervals, settings, sounds
+}
+
+private struct IntervalsView: View {
+    @EnvironmentObject var timer: TBTimer
+    private var minStr = NSLocalizedString("IntervalsView.min", comment: "min")
+
+    var body: some View {
+        VStack {
+            Stepper(value: $timer.workIntervalLength, in: 1 ... 60) {
+                HStack {
+                    Text(NSLocalizedString("IntervalsView.workIntervalLength.label",
+                                           comment: "Work interval label"))
+                        .frame(maxWidth: .infinity, alignment: .leading)
+                    Text(String.localizedStringWithFormat(minStr, timer.workIntervalLength))
+                }
+            }
+            Stepper(value: $timer.shortRestIntervalLength, in: 1 ... 60) {
+                HStack {
+                    Text(NSLocalizedString("IntervalsView.shortRestIntervalLength.label",
+                                           comment: "Short rest interval label"))
+                        .frame(maxWidth: .infinity, alignment: .leading)
+                    Text(String.localizedStringWithFormat(minStr, timer.shortRestIntervalLength))
+                }
+            }
+            Stepper(value: $timer.longRestIntervalLength, in: 1 ... 60) {
+                HStack {
+                    Text(NSLocalizedString("IntervalsView.longRestIntervalLength.label",
+                                           comment: "Long rest interval label"))
+                        .frame(maxWidth: .infinity, alignment: .leading)
+                    Text(String.localizedStringWithFormat(minStr, timer.longRestIntervalLength))
+                }
+            }
+            .help(NSLocalizedString("IntervalsView.longRestIntervalLength.help",
+                                    comment: "Long rest interval hint"))
+            Stepper(value: $timer.workIntervalsInSet, in: 1 ... 10) {
+                HStack {
+                    Text(NSLocalizedString("IntervalsView.workIntervalsInSet.label",
+                                           comment: "Work intervals in a set label"))
+                        .frame(maxWidth: .infinity, alignment: .leading)
+                    Text("\(timer.workIntervalsInSet)")
+                }
+            }
+            .help(NSLocalizedString("IntervalsView.workIntervalsInSet.help",
+                                    comment: "Work intervals in set hint"))
+            Spacer().frame(minHeight: 0)
+        }
+        .padding(4)
+    }
+}
+
+
+
+
+
+
+
+
+
+
+
+#if DEBUG
+    func debugSize(proxy: GeometryProxy) -> some View {
+        print("Optimal popover size:", proxy.size)
+        return Color.clear
+    }
+#endif
