+++ TomatoBar/Views/TBPopoverView.swift
@@ -1,0 +1,106 @@
+//
+//  TBPopoverView.swift
+//  TomatoBar
+//
+//  Created by Pierre Oosthuizen on 2026/04/24.
+//
+
+
+struct TBPopoverView: View {
+    @ObservedObject var timer = TBTimer()
+    @State private var buttonHovered = false
+    @State private var activeChildView = ChildView.intervals
+
+    private var startLabel = NSLocalizedString("TBPopoverView.start.label", comment: "Start label")
+    private var stopLabel = NSLocalizedString("TBPopoverView.stop.label", comment: "Stop label")
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 8) {
+            Button {
+                timer.startStop()
+                TBStatusItem.shared.closePopover(nil)
+            } label: {
+                Text(timer.timer != nil ?
+                     (buttonHovered ? stopLabel : timer.timeLeftString) :
+                        startLabel)
+                    /*
+                      When appearance is set to "Dark" and accent color is set to "Graphite"
+                      "defaultAction" button label's color is set to the same color as the
+                      button, making the button look blank. #24
+                     */
+                    .foregroundColor(Color.white)
+                    .font(.system(.body).monospacedDigit())
+                    .frame(maxWidth: .infinity)
+            }
+            .onHover { over in
+                buttonHovered = over
+            }
+            .controlSize(.large)
+            .keyboardShortcut(.defaultAction)
+
+            Picker("", selection: $activeChildView) {
+                Text(NSLocalizedString("TBPopoverView.intervals.label",
+                                       comment: "Intervals label")).tag(ChildView.intervals)
+                Text(NSLocalizedString("TBPopoverView.settings.label",
+                                       comment: "Settings label")).tag(ChildView.settings)
+                Text(NSLocalizedString("TBPopoverView.sounds.label",
+                                       comment: "Sounds label")).tag(ChildView.sounds)
+            }
+            .labelsHidden()
+            .frame(maxWidth: .infinity)
+            .pickerStyle(.segmented)
+
+            GroupBox {
+                switch activeChildView {
+                case .intervals:
+                    IntervalsView().environmentObject(timer)
+                case .settings:
+                    SettingsView().environmentObject(timer)
+                case .sounds:
+                    SoundsView().environmentObject(timer.player)
+                }
+            }
+
+            Group {
+                Button {
+                    NSApp.activate(ignoringOtherApps: true)
+                    NSApp.orderFrontStandardAboutPanel()
+                } label: {
+                    Text(NSLocalizedString("TBPopoverView.about.label",
+                                           comment: "About label"))
+                    Spacer()
+                    Text("⌘ A").foregroundColor(Color.gray)
+                }
+                .buttonStyle(.plain)
+                .keyboardShortcut("a")
+                Button {
+                    NSApplication.shared.terminate(self)
+                } label: {
+                    Text(NSLocalizedString("TBPopoverView.quit.label",
+                                           comment: "Quit label"))
+                    Spacer()
+                    Text("⌘ Q").foregroundColor(Color.gray)
+                }
+                .buttonStyle(.plain)
+                .keyboardShortcut("q")
+            }
+        }
+        #if DEBUG
+            /*
+             After several hours of Googling and trying various StackOverflow
+             recipes I still haven't figured a reliable way to auto resize
+             popover to fit all it's contents (pull requests are welcome!).
+             The following code block is used to determine the optimal
+             geometry of the popover.
+             */
+            .overlay(
+                GeometryReader { proxy in
+                    debugSize(proxy: proxy)
+                }
+            )
+        #endif
+            /* Use values from GeometryReader */
+//            .frame(width: 240, height: 276)
+            .padding(12)
+    }
+}
